+++ Microsoft.Azure.Cosmos.Samples/CodeSamples/appSettings.json
@@ -1,4 +1,4 @@
 {
-  "EndPointUrl": "https://localhost:8081",
-  "AuthorizationKey": "Super secret key"
+  "EndPointUrl": "https://jawilleytemp.documents.azure.com:443/",
+  "AuthorizationKey": "GboKIMLEDVsipUwUx9qVo9k7zACZqhOG3Kh9PkWPeZK2FuSAAEeG6R3T3TinPKh3upxV4OYY4uw6tJI7whtfDA=="
 }
